chore: 🧹 clean up project structure
