--- mobile/lib/app/routes/app_pages.dart
@@ -1,55 +1,0 @@
-import 'package:get/get.dart';
-
-import '../modules/attendance/bindings/attendance_binding.dart';
-import '../modules/attendance/views/attendance_view.dart';
-import '../modules/bottomNavBar/bindings/bottom_nav_bar_binding.dart';
-import '../modules/bottomNavBar/views/bottom_nav_bar_view.dart';
-import '../modules/home/bindings/home_binding.dart';
-import '../modules/home/views/home_view.dart';
-import '../modules/login/bindings/login_binding.dart';
-import '../modules/login/views/login_view.dart';
-import '../modules/profile/bindings/profile_binding.dart';
-import '../modules/profile/views/profile_view.dart';
-import '../modules/splash/bindings/splash_binding.dart';
-import '../modules/splash/views/splash_view.dart';
-
-part 'app_routes.dart';
-
-class AppPages {
-  AppPages._();
-
-  static const INITIAL = Routes.SPLASH;
-
-  static final routes = [
-    GetPage(
-      name: _Paths.HOME,
-      page: () => const HomeView(),
-      binding: HomeBinding(),
-    ),
-    GetPage(
-      name: _Paths.SPLASH,
-      page: () => const SplashView(),
-      binding: SplashBinding(),
-    ),
-    GetPage(
-      name: _Paths.LOGIN,
-      page: () => const LoginView(),
-      binding: LoginBinding(),
-    ),
-    GetPage(
-      name: _Paths.PROFILE,
-      page: () => const ProfileView(),
-      binding: ProfileBinding(),
-    ),
-    GetPage(
-      name: _Paths.BOTTOM_NAV_BAR,
-      page: () => const BottomNavBarView(),
-      binding: BottomNavBarBinding(),
-    ),
-    GetPage(
-      name: _Paths.ATTENDANCE,
-      page: () => const AttendanceView(),
-      binding: AttendanceBinding(),
-    ),
-  ];
-}
