+++ mobile/lib/core/routes/routes.dart
@@ -1,0 +1,1 @@
+export './app_pages.dart';
