+++ mobile/lib/main.dart
@@ -1,8 +1,7 @@
+import 'package:bantu_pengusaha/core/routes/routes.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:oktoast/oktoast.dart';
-
-import 'app/routes/app_pages.dart';
 
 void main() {
   runApp(
