+++ mobile/lib/modules/attendance/attendance.dart
@@ -1,0 +1,3 @@
+export './bindings/attendance_binding.dart';
+export './controllers/attendance_controller.dart';
+export './views/attendance_view.dart';
