+++ mobile/lib/modules/bottom_nav_bar/bindings/bottom_nav_bar_binding.dart
