+++ mobile/lib/modules/bottom_nav_bar/bottom_nav_bar.dart
@@ -1,0 +1,3 @@
+export './bindings/bottom_nav_bar_binding.dart';
+export './controllers/bottom_nav_bar_controller.dart';
+export './views/bottom_nav_bar_view.dart';
