+++ mobile/lib/modules/bottom_nav_bar/controllers/bottom_nav_bar_controller.dart
