+++ mobile/lib/modules/home/home.dart
@@ -1,0 +1,3 @@
+export './bindings/home_binding.dart';
+export './controllers/home_controller.dart';
+export './views/home_view.dart';
