+++ mobile/lib/modules/login/login.dart
@@ -1,0 +1,3 @@
+export './bindings/login_binding.dart';
+export './controllers/login_controller.dart';
+export './views/login_view.dart';
