+++ mobile/lib/modules/modules.dart
@@ -1,0 +1,6 @@
+export './attendance/attendance.dart';
+export './bottom_nav_bar/bottom_nav_bar.dart';
+export './home/home.dart';
+export './login/login.dart';
+export './profile/profie.dart';
+export './splash/splash.dart';
