+++ mobile/lib/modules/profile/profie.dart
@@ -1,0 +1,3 @@
+export './bindings/profile_binding.dart';
+export './controllers/profile_controller.dart';
+export './views/profile_view.dart';
