+++ mobile/lib/modules/splash/splash.dart
@@ -1,0 +1,3 @@
+export './bindings/splash_binding.dart';
+export './controllers/splash_controller.dart';
+export './views/splash_view.dart';
