--- mobile/lib/utils/extensions.dart
@@ -1,1 +1,0 @@
-
