+++ mobile/lib/utils/utils.dart
@@ -1,2 +1,1 @@
 export 'logger.dart';
-export 'extensions.dart';
